move example/server/*.go to example/server/exampleop/
